Re-init git repo.
This is a fork of a checkout representing November 2016 work by the
original PyGirl authors. GitHub has deleted the original repository and
I only had this source due to Nix caching, which is a snapshot of a
working tree with no git history.

If I can find more complete history, then I will rebase every following
change to incorporate that history.