Remove tests.
Yes, I know that this is dramatic, but fundamentally this test suite is
far too brittle for a working emulator -- we should be able to describe
non-regressed functionality in terms of specific test ROMs, rather than
unit-testing internal interfaces.

Because, as we'll see, the internal interfaces can be improved
dramatically, and I'm not getting paid to write unit tests.