video: Write lines, not pixels, to the framebuffer.
This change removes a bottleneck caused by the overhead of repeatedly
performing per-pixel logic that can mostly be performed per-line
instead. Don't worry too much about the RPython encoding; just notice
that bytearray is mutable str, and the messy index-driven copying is
just a memcpy idiom that Python 2.7 didn't have.

The new bottleneck is SDL_FillRect(), which takes up 41% of time on its
own time and 64% of total time when including its FFI overhead. This is
because we are calling it once per pixel redraw, which is unacceptable.