Much faster screen updates.
We used to spend all of our time in SDL_FillRect(), which essentially is
a 2D loop that fills in a rectangle by visiting a Cartesian product of
pixel coordinates. It's not a complex algorithm. However, the overhead
of FFI calls was staggering; we were spending 15-20% of time either
preparing or recovering from FFI.

Now, we do something much simpler conceptually: we allocate at least two
screen surfaces by passing SDL.DOUBLEBUF, and we write an entirely new
screen every frame by inlining a version of that 2D loop within another
2D loop that visits every coordinate in GB screen space. We're
effectively writing the same pixels, but we now only have to make O(1)
FFI calls to lock, unlock, and flip the surfaces.

While normally four nested loops would be disastrous, it's not a problem
here since the inner two loops are tiny. Still, I'd like to be able to
use library code to directly scale a surface, as in SDL 2, and it's a
little frustrating to lose a few hours of hacking to pointer arithmetic
for lack of a library procedure.

The actual amount of CPU consumed is not much different, at least on my
laptop, but a few demos are noticably faster. I'm guessing that we were
generating (real-time) lag frames but capped in CPU consumption by
vsync. After this change, perf says that our top two frames are
VideoDriverImplementation.draw_pixels() followed by CPU.emulate(), which
means that any time we saved in this change could result in less lag.