Wire up sound with SDL.
This was a little tedious and a little confusing. Also, sound doesn't
actually work yet.

This does translate, though. It also replaces the old sampling path,
which filled an entire buffer at once and stored it for eventual output,
with a more reasonable real-time path which actuates the sound chip on
its own without interfering with the rest of the emulator. Along the
way, the Sound and SoundDriver classes were disentangled, many side
methods were deleted, and several constructors were simplified.