sound: Run the sound chip from the SDL callback.
In emulator lore, it's known that either audio or video must be run at
the rate of the emulating machine's interrupts, and the other must be
dynamically adjusted to stay within one frame of the emulated machine.
This is a rather non-negotiable situation, in that the only tradeoff to
make is to dynamically adjust *both* audio and video.

So, like many emulators, we will run the relatively slow SPU in the
middle of the audio callback. In the typical case, at 44.1 KHz, with 512
samples to put in the buffer, we will run the callback at 86.1328125 Hz
and the SPU at 256 Hz. This means that the callback will ask the SPU to
update itself 2.972154195011338 times per run... which is 1% off from 3
and going to cause an audible buzz if not compensated.

The compensation looks like a PLL with Bresenham's algorithm for laying
out the coefficients in the train. I fully admit that I've forgotten the
optimized way to do this and open-coded a fairly messy loop instead. I
also cheated and used floats; Bresenham is dishonored by my failure to
stick to ints, but I don't think it'll be a performance problem. The
train has 64 == 2 ** 6 coefficients to grant an extra 6 powers of two to
44100, which has 2 itself. 2 + 6 == 8 == log2(256), so this is just
barely enough divisors to allow for perfectly running 44.1 KHz audio. Of
course 48 KHz already had 7 powers of two and so this is delicious
overkill for that case.

Audio still doesn't sound great. There's lots of slow-beating glitches
in values that I suspect are caused by the much faster CPU racing the
now-much-slower SPU, which operates in a separate thread.