sound: Lock when writing/reading registers.
I'm not sure exactly how much this helps, but it *does* help; a few
demos are noticably cleaner, to the point where drums and sustained
melodic lines are perceptible. There's a serious deficit of volume
control, lots of screech, and a general disappointment, but it's getting
there. Slowly but surely.